Primera Version de la pagina de competiciones
