--- .astro/settings.json
@@ -1,8 +1,0 @@
-{
-	"_variables": {
-		"lastUpdateCheck": 1736331030424
-	},
-	"devToolbar": {
-		"enabled": false
-	}
-}
