+++ package.json
@@ -14,7 +14,7 @@
     "@astrojs/sitemap": "^3.1.1",
     "@astrojs/tailwind": "^5.0.0",
     "@fontsource-variable/inter": "^5.0.8",
-    "astro": "^4.16.7",
+    "astro": "^3.6.4",
     "astro-compress": "^2.0.15",
     "sharp": "^0.33.2",
     "swiper": "^11.0.4",
